added timing to do_job

+++ matrix.h
@@ -6,4 +6,4 @@ void multiply_matrices(int rows1, int cols1, int *matrix1,
 
 void display_matrix(int rows, int cols, int *matrix);
 
-int do_job(int rows1, int cols1, int cols2, int forever);
+float do_job(int rows1, int cols1, int cols2, int forever);
